# Benchmark and optimize PushRequest