While this is a seemingly innocent code path, it shows up in pprof's a lot - like 5-10%+ of CPU and allocs.

```
name           old time/op    new time/op    delta
PushRequest-6    6.15ms ± 1%    5.42ms ± 8%  -11.87%  (p=0.002 n=6+6)

name           old alloc/op   new alloc/op   delta
PushRequest-6    4.24MB ± 0%    3.48MB ± 0%  -17.94%  (p=0.002 n=6+6)

name           old allocs/op  new allocs/op  delta
PushRequest-6     85.1k ± 0%     60.1k ± 0%  -29.39%  (p=0.000 n=5+6)
```

This will get faster with the fixes in https://github.com/census-instrumentation/opencensus-go/issues/1265 as well